add --mount-home-dir, mount temp dir as home dir by default
the --mount-home-dir option is nice for caching dot directories across builds. bind-mounting the home dir also avoids
most problems with mangling bind-mounts in docker requests from inside the container (using --docker-proxy) when we
cannot detect the container's rootfs dir on the host (possibly because the host docker is not using the 'overlay' driver
for mounts).